+++ lib/token_field/simple_form/token_input.rb
@@ -1,0 +1,13 @@
+require "simple_form/inputs/base"
+
+class TokenInput < SimpleForm::Inputs::Base
+  #
+  # one:
+  # <%= f.input :category_id, :as => :token %>
+  #
+  # many:
+  # <%= f.input :category_ids, :as => :token %>
+  def input
+    @builder.token_field(attribute_name, options)
+  end
+end