+++ spec/dummy/app/models/item.rb
@@ -1,0 +1,5 @@
+class Item < ActiveRecord::Base
+  attr_accessible :category_id, :name
+
+  belongs_to :category
+end