+++ changelog.md
@@ -1,5 +1,12 @@
+# 0.0.2
+## added
+* support for simple_form with TokenInput class
+* tests for simple_form
+## changed
+* readme file
+
 # 0.0.1
-# added
+## added
 * init version of gem
 * view helper token_input
 * testing helper fill_in_token